Add -d option

+++ main.go
@@ -1,6 +1,7 @@
 package main
 
 import (
+	"flag"
 	"fmt"
 	"image/jpeg"
 	"image/png"
@@ -9,18 +10,24 @@ import (
 	"strings"
 )
 
+var (
+	delOpt = flag.Bool("d", false, "Add this option if you want to delete the target file after exif deletion. In that case, the file with `_exif-deleted.JPG` will not be generated.")
+)
+
 func usage() {
-	fmt.Println("Usage: go-delete-exif <input image path(JPEG only)>")
+	fmt.Println("Usage: go-delete-exif <-d (optional)> <input image path(JPEG only)>")
 }
 
 func main() {
-	if len(os.Args) != 2 {
+	if len(os.Args) != 2 && len(os.Args) != 3 {
 		fmt.Println("Error: You must specify the image (JPEG) from which you want to delete Exif.")
 		usage()
 		os.Exit(0)
 	}
 
-	imgPath := os.Args[1]
+	flag.Parse()
+
+	imgPath := os.Args[len(os.Args)-1]
 	_, err := os.Stat(imgPath)
 	if err != nil {
 		fmt.Println("Error: Invalid specify filepath")
@@ -54,22 +61,47 @@ func main() {
 	// convert png and delete exif
 	png.Encode(tmpImg, img)
 
-	outImgName := strings.Split(filepath.Base(imgPath), ".")[0] + "_exif-deleted.JPG"
-	outputPath := filepath.Join(filepath.Dir(imgPath), outImgName)
-	outImg, err := os.Create(outputPath)
-	if err != nil {
-		fmt.Printf("Error: %v\n", err)
-		os.Exit(1)
+	if *delOpt {
+		// Delete target file
+		if err := os.Remove(imgPath); err != nil {
+			fmt.Printf("Error: %v\n", err)
+			os.Exit(1)
+		}
+
+		outImg, err := os.Create(imgPath)
+		if err != nil {
+			fmt.Printf("Error: %v\n", err)
+			os.Exit(1)
+		}
+		defer outImg.Close()
+
+		jpeg.Encode(outImg, img, nil)
+
+		// Delete tmp file
+		if err := os.Remove(tmpPngPath); err != nil {
+			fmt.Printf("Error: %v\n", err)
+			os.Exit(1)
+		}
+
+		fmt.Printf("Success: delete exif ===> %s\n", imgPath)
+	} else {
+		outImgName := strings.Split(filepath.Base(imgPath), ".")[0] + "_exif-deleted.JPG"
+		outputPath := filepath.Join(filepath.Dir(imgPath), outImgName)
+		outImg, err := os.Create(outputPath)
+		if err != nil {
+			fmt.Printf("Error: %v\n", err)
+			os.Exit(1)
+		}
+		defer outImg.Close()
+
+		jpeg.Encode(outImg, img, nil)
+
+		// Delete tmp file
+		if err := os.Remove(tmpPngPath); err != nil {
+			fmt.Printf("Error: %v\n", err)
+			os.Exit(1)
+		}
+
+		fmt.Printf("Success: delete exif ===> %s\n", outputPath)
 	}
-	defer outImg.Close()
-
-	jpeg.Encode(outImg, img, nil)
-
-	// Delete tmp file
-	if err := os.Remove(tmpPngPath); err != nil {
-		fmt.Printf("Error: %v\n", err)
-		os.Exit(1)
-	}
-
-	fmt.Printf("Success: delete exif ===> %s\n", outputPath)
 }
